apps/server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
websockets>=12.0
pydantic>=2.8.0
pydantic-settings>=2.3.0
dashscope>=1.23.9
redis>=5.0.0
pypdf>=4.2.0
python-docx>=1.1.2
